Removing require from www putting back sequelize in www sequelize comes from config.js added config.js
