+++ templates/js/www.ejs
@@ -4,30 +4,40 @@
  * Module dependencies.
  */
 
-var app = require('../app');
-var debug = require('debug')('<%- name %>:server');
-var http = require('http');
+import debugPkg from 'debug';
+import http from 'http';
+import app from '../app';
+<% if (db) { -%>
+import db from '../models';
+<% } -%>
+
+const debug = debugPkg('<%- name %>:server');
 
 /**
  * Get port from environment and store in Express.
  */
 
-var port = normalizePort(process.env.PORT || '3000');
+const port = normalizePort(process.env.PORT || '3000');
 app.set('port', port);
 
 /**
  * Create HTTP server.
  */
 
-var server = http.createServer(app);
+const server = http.createServer(app);
 
 /**
  * Listen on provided port, on all network interfaces.
  */
 
+ <% if (db) { %>
+  // database setup
+  <%- db %>;
+<%} else { %>
 server.listen(port, function() {
   console.log("App listening on PORT " + port);
 });
+<% } %>
 
 server.on('error', onError);
 server.on('listening', onListening);
